Attach content-type directly to gridfs object
GridFS offers a special attribute to attach a content-type to a file.
This commit introduces an extra arugment to add artifact to put the
content-type into. If the content-type is not set manually it is
automatically detected. However, this behavior makes more sense then
reserving a field in the artifacts metadata for the content-type. This
change has only an effect on the MongoObserver.