Fixes #442: config_hook overwritten by config_updates
The updates contained in the return dictionary of a `config_hook` were overwritten if a corresponding
entry existed in `config_updates` as passed e.g. from the command line. This commit removes the line
that is responsible for this behavior, as well as adding a test case and a slightly modified documentation.