Terminate tee subprocesses before waiting
Send termination signals to tee_stdout and tee_stderr, before waiting.
This is needed because on some systems there seems to be deadlock
issues related to using multiprocessing combined with fd capture.
https://github.com/IDSIA/sacred/issues/705#issuecomment-564918590
shows an example of this.

This change reliably fixes the problem on the system I am using,
and hopefully for others as well. There should be no ill effects
from terminating the subprocesses because this is done after the
stdin pipe is closed.

Fixes https://github.com/IDSIA/sacred/issues/705